+++ example/factories.ts
@@ -2,18 +2,25 @@ import { lookup } from "mime-types";
 import {
   createApiResponse,
   createResultHandler,
-  defaultEndpointsFactory,
+  defaultResultHandler,
   EndpointsFactory,
   z,
 } from "../src";
+import { config } from "./config";
 import { authMiddleware } from "./middlewares";
 import fs from "fs";
 
-export const keyAndTokenAuthenticatedEndpointsFactory =
-  defaultEndpointsFactory.addMiddleware(authMiddleware);
+export const taggedEndpointsFactory = new EndpointsFactory({
+  resultHandler: defaultResultHandler,
+  config,
+});
 
-export const fileSendingEndpointsFactory = new EndpointsFactory(
-  createResultHandler({
+export const keyAndTokenAuthenticatedEndpointsFactory =
+  taggedEndpointsFactory.addMiddleware(authMiddleware);
+
+export const fileSendingEndpointsFactory = new EndpointsFactory({
+  config,
+  resultHandler: createResultHandler({
     getPositiveResponse: () =>
       createApiResponse(z.string(), lookup("svg") || "image/svg+xml"),
     getNegativeResponse: () =>
@@ -29,11 +36,12 @@ export const fileSendingEndpointsFactory = new EndpointsFactory(
         response.status(400).send("Data is missing");
       }
     },
-  })
-);
+  }),
+});
 
-export const fileStreamingEndpointsFactory = new EndpointsFactory(
-  createResultHandler({
+export const fileStreamingEndpointsFactory = new EndpointsFactory({
+  config,
+  resultHandler: createResultHandler({
     getPositiveResponse: () => createApiResponse(z.file().binary(), "image/*"),
     getNegativeResponse: () =>
       createApiResponse(z.string(), lookup("txt") || "text/plain"),
@@ -50,5 +58,5 @@ export const fileStreamingEndpointsFactory = new EndpointsFactory(
         response.status(400).send("Filename is missing");
       }
     },
-  })
-);
+  }),
+});