+++ src/endpoints-factory.ts
@@ -8,6 +8,7 @@ import {
   IOSchema,
   ProbableIntersection,
 } from "./common-helpers";
+import { CommonConfig } from "./config-type";
 import { Endpoint, Handler } from "./endpoint";
 import { Method, MethodsDefinition } from "./method";
 import {
@@ -15,7 +16,6 @@ import {
   createMiddleware,
   ExpressMiddleware,
   ExpressMiddlewareFeatures,
-  MiddlewareCreationProps,
   MiddlewareDefinition,
 } from "./middleware";
 import { mimeJson, mimeMultipart } from "./mime";
@@ -30,37 +30,59 @@ type BuildProps<
   MIN extends IOSchema<"strip"> | null,
   OPT extends FlatObject,
   M extends Method,
-  SCO extends string
+  SCO extends string,
+  TAG extends string
 > = {
   input: IN;
   output: OUT;
   handler: Handler<z.output<ProbableIntersection<MIN, IN>>, z.input<OUT>, OPT>;
   description?: string;
-  scopes?: SCO[];
-} & MethodsDefinition<M>;
+  shortDescription?: string;
+} & ({ scopes?: SCO[] } | { scope?: SCO }) &
+  ({ tags?: TAG[] } | { tag?: TAG }) &
+  MethodsDefinition<M>;
 
 export class EndpointsFactory<
   POS extends ApiResponse,
   NEG extends ApiResponse,
   IN extends IOSchema<"strip"> | null = null,
   OUT extends FlatObject = {},
-  SCO extends string = string
+  SCO extends string = string,
+  TAG extends string = string
 > {
+  protected resultHandler: ResultHandlerDefinition<POS, NEG>;
   protected middlewares: AnyMiddlewareDef[] = [];
 
-  constructor(protected resultHandler: ResultHandlerDefinition<POS, NEG>) {}
+  constructor(resultHandler: ResultHandlerDefinition<POS, NEG>);
+  /** @desc Consider using the "config" prop with the "tags" option to enforce constraints on tagging the endpoints */
+  constructor(params: {
+    resultHandler: ResultHandlerDefinition<POS, NEG>;
+    config?: CommonConfig<TAG>;
+  });
+  constructor(
+    subject:
+      | ResultHandlerDefinition<POS, NEG>
+      | {
+          resultHandler: ResultHandlerDefinition<POS, NEG>;
+          config?: CommonConfig<TAG>;
+        }
+  ) {
+    this.resultHandler =
+      "resultHandler" in subject ? subject.resultHandler : subject;
+  }
 
   static #create<
     CPOS extends ApiResponse,
     CNEG extends ApiResponse,
     CIN extends IOSchema<"strip"> | null,
     COUT extends FlatObject,
-    CSCO extends string
+    CSCO extends string,
+    CTAG extends string
   >(
     middlewares: AnyMiddlewareDef[],
     resultHandler: ResultHandlerDefinition<CPOS, CNEG>
   ) {
-    const factory = new EndpointsFactory<CPOS, CNEG, CIN, COUT, CSCO>(
+    const factory = new EndpointsFactory<CPOS, CNEG, CIN, COUT, CSCO, CTAG>(
       resultHandler
     );
     factory.middlewares = middlewares;
@@ -71,52 +93,15 @@ export class EndpointsFactory<
     AIN extends IOSchema<"strip">,
     AOUT extends FlatObject,
     ASCO extends string
-  >(
-    definition: MiddlewareDefinition<AIN, OUT, AOUT, ASCO>
-  ): EndpointsFactory<
-    POS,
-    NEG,
-    ProbableIntersection<IN, AIN>,
-    OUT & AOUT,
-    SCO & ASCO
-  >;
-
-  /** @deprecated please use createMiddleware() for the argument */
-  public addMiddleware<
-    AIN extends IOSchema<"strip">,
-    AOUT extends FlatObject,
-    ASCO extends string
-  >(
-    props: MiddlewareCreationProps<AIN, OUT, AOUT, ASCO>
-  ): EndpointsFactory<
-    POS,
-    NEG,
-    ProbableIntersection<IN, AIN>,
-    OUT & AOUT,
-    SCO & ASCO
-  >;
-
-  public addMiddleware<
-    AIN extends IOSchema<"strip">,
-    AOUT extends FlatObject,
-    ASCO extends string
-  >(
-    subject:
-      | MiddlewareDefinition<AIN, OUT, AOUT, ASCO>
-      | MiddlewareCreationProps<AIN, OUT, AOUT, ASCO>
-  ) {
+  >(subject: MiddlewareDefinition<AIN, OUT, AOUT, ASCO>) {
     return EndpointsFactory.#create<
       POS,
       NEG,
       ProbableIntersection<IN, AIN>,
       OUT & AOUT,
-      SCO & ASCO
-    >(
-      this.middlewares.concat(
-        "type" in subject ? subject : createMiddleware(subject)
-      ),
-      this.resultHandler
-    );
+      SCO & ASCO,
+      TAG
+    >(this.middlewares.concat(subject), this.resultHandler);
   }
 
   public use = this.addExpressMiddleware;
@@ -145,14 +130,14 @@ export class EndpointsFactory<
           middleware(request as R, response as S, next);
         }),
     };
-    return EndpointsFactory.#create<POS, NEG, IN, OUT & AOUT, SCO>(
+    return EndpointsFactory.#create<POS, NEG, IN, OUT & AOUT, SCO, TAG>(
       this.middlewares.concat(definition),
       this.resultHandler
     );
   }
 
   public addOptions<AOUT extends FlatObject>(options: AOUT) {
-    return EndpointsFactory.#create<POS, NEG, IN, OUT & AOUT, SCO>(
+    return EndpointsFactory.#create<POS, NEG, IN, OUT & AOUT, SCO, TAG>(
       this.middlewares.concat(
         createMiddleware({
           input: z.object({}),
@@ -166,22 +151,21 @@ export class EndpointsFactory<
   public build<BIN extends IOSchema, BOUT extends IOSchema, M extends Method>({
     input,
     handler,
-    description,
     output: outputSchema,
     ...rest
-  }: BuildProps<BIN, BOUT, IN, OUT, M, SCO>): Endpoint<
+  }: BuildProps<BIN, BOUT, IN, OUT, M, SCO, TAG>): Endpoint<
     ProbableIntersection<IN, BIN>,
     BOUT,
     OUT,
     M,
     POS,
     NEG,
-    SCO
+    SCO,
+    TAG
   > {
     const { middlewares, resultHandler } = this;
     return new Endpoint({
       handler,
-      description,
       middlewares,
       outputSchema,
       resultHandler,