+++ tests/unit/endpoint.spec.ts
@@ -415,15 +415,16 @@ describe("Endpoint", () => {
   describe("Issue #514: Express native middlewares for OPTIONS request", () => {
     test("should skip proprietary ones", async () => {
       const endpoint = new EndpointsFactory(defaultResultHandler)
-        .addMiddleware({
-          // testing also the backward compatibility (without createMiddleware)
-          input: z.object({
-            shouldNotBeHere: z.boolean(),
-          }),
-          middleware: async () => {
-            throw new Error("Should not be here");
-          },
-        })
+        .addMiddleware(
+          createMiddleware({
+            input: z.object({
+              shouldNotBeHere: z.boolean(),
+            }),
+            middleware: async () => {
+              throw new Error("Should not be here");
+            },
+          })
+        )
         .addExpressMiddleware((req, res, next) => {
           res.set("X-Custom-Header", "test");
           next();
@@ -447,6 +448,100 @@ describe("Endpoint", () => {
       expect(responseMock.status).toHaveBeenCalledWith(200);
       expect(responseMock.json).toHaveBeenCalledTimes(0);
       expect(responseMock.set).toHaveBeenCalledWith("X-Custom-Header", "test");
+    });
+  });
+
+  describe("Issue #585: Handling non-Error exceptions", () => {
+    test("thrown in #parseOutput()", async () => {
+      const factory = new EndpointsFactory(defaultResultHandler);
+      const endpoint = factory.build({
+        method: "post",
+        input: z.object({}),
+        output: z.object({
+          test: z.number().transform(() => {
+            // eslint-disable-next-line @typescript-eslint/no-throw-literal
+            throw "Something unexpected";
+          }),
+        }),
+        handler: async () => ({
+          test: 123,
+        }),
+      });
+      const { responseMock, loggerMock } = await testEndpoint({
+        endpoint,
+      });
+      expect(loggerMock.error).toBeCalledTimes(1);
+      expect(responseMock.status).toBeCalledWith(500);
+      expect(responseMock.json).toBeCalledWith({
+        status: "error",
+        error: {
+          message: "Something unexpected",
+        },
+      });
+    });
+
+    test("thrown in #handleResult()", async () => {
+      const factory = new EndpointsFactory(
+        createResultHandler({
+          getPositiveResponse: () => createApiResponse(z.object({})),
+          getNegativeResponse: () => createApiResponse(z.object({})),
+          handler: () => {
+            // eslint-disable-next-line @typescript-eslint/no-throw-literal
+            throw "Something unexpected happened";
+          },
+        })
+      );
+      const endpoint = factory.build({
+        method: "get",
+        input: z.object({}),
+        output: z.object({
+          test: z.string(),
+        }),
+        handler: async () => ({ test: "OK" }),
+      });
+      const { loggerMock, responseMock } = await testEndpoint({ endpoint });
+      expect(loggerMock.error).toBeCalledTimes(1);
+      expect(loggerMock.error.mock.calls[0][0]).toBe(
+        "Result handler failure: Something unexpected happened."
+      );
+      expect(responseMock.status).toBeCalledTimes(1);
+      expect(responseMock.status.mock.calls[0][0]).toBe(500);
+      expect(responseMock.json).toBeCalledTimes(0);
+      expect(responseMock.end).toBeCalledTimes(1);
+      expect(responseMock.end.mock.calls[0][0]).toBe(
+        "An error occurred while serving the result: Something unexpected happened."
+      );
+    });
+
+    test("thrown in middleware and caught in execute()", async () => {
+      const factory = new EndpointsFactory(defaultResultHandler).addMiddleware(
+        createMiddleware({
+          input: z.object({}),
+          middleware: async () => {
+            // eslint-disable-next-line @typescript-eslint/no-throw-literal
+            throw "Something went wrong";
+          },
+        })
+      );
+      const endpoint = factory.build({
+        methods: ["post"],
+        input: z.object({}),
+        output: z.object({}),
+        handler: async () => ({}),
+      });
+      const { responseMock, loggerMock } = await testEndpoint({
+        endpoint,
+        requestProps: {
+          method: "POST",
+          body: {},
+        },
+      });
+      expect(loggerMock.error).toBeCalledTimes(1);
+      expect(responseMock.status).toBeCalledWith(500);
+      expect(responseMock.json).toBeCalledWith({
+        status: "error",
+        error: { message: "Something went wrong" },
+      });
     });
   });
 });