+++ CHANGELOG.md
@@ -1,6 +1,139 @@
 # Changelog
 
+## Version 8
+
+### v8.2.0
+
+- Feature #637: endpoint short description (summary).
+  - Added the ability to assign a `shortDescription` to endpoints.
+  - One sentence of no more than 50 characters is implied.
+  - This text is substituted into the `summary` property of the generated documentation.
+  - Visualizers of the generated OpenAPI documentation nicely display this text on the same line as the endpoint path.
+  - If a `shortDescription` is not specified, but a regular `description` is, then by default the `summary` will be
+    generated from the `description` by trimming.
+  - You can optionally disable this behavior with the new option `hasSummaryFromDescription` of the `OpenAPI` generator.
+
+```typescript
+const exampleEndpoint = yourEndpointsFactory.build({
+  // ...
+  description: "The detailed explanaition on what this endpoint does.",
+  shortDescription: "Retrieves the user.",
+});
+```
+
+### v8.1.0
+
+- Feature #571: tagging the endpoints.
+  - Good news dear community! You can now tag your endpoints using the new properties of the `.build()` method
+    of the `EndpointsFactory`.
+  - For your convenience and for the sake of Semantics, there are singular and plural properties: `tag` and `tags`.
+  - By default, these properties allow any string, so in order to enforce restrictions and achieve the consistency
+    across all endpoints, the possible tags should be declared in the configuration first and also a brand
+    new `EndpointsFactory` instantiation approach is required.
+  - The configuration has got a new `tags` property for declaring possible tags and their descriptions.
+  - Tags are an important part of the generated documentation for the OpenAPI standard.
+- The property `scopes` (introduced in v7.9.0) has got its singular variation `scope`.
+
+```typescript
+import {
+  createConfig,
+  EndpointsFactory,
+  defaultResultHandler,
+} from "express-zod-api";
+
+const config = createConfig({
+  // ..., use the simple or the advanced syntax:
+  tags: {
+    users: "Everything about the users",
+    files: {
+      description: "Everything about the files processing",
+      url: "https://example.com",
+    },
+  },
+});
+
+// instead of defaultEndpointsFactory use the following approach:
+const taggedEndpointsFactory = new EndpointsFactory({
+  resultHandler: defaultResultHandler, // or use your custom one
+  config,
+});
+
+const exampleEndpoint = taggedEndpointsFactory.build({
+  // ...
+  tag: "users", // or tags: ["users", "files"]
+});
+```
+
+### v8.0.2
+
+- `express` version is 4.18.2.
+- `openapi3-ts` version is 3.1.0.
+
+### v8.0.1
+
+- `zod` version is 3.19.1.
+
+### v8.0.0
+
+- `winston` version is 3.8.2.
+- This version is based on v8.0.0-beta3 and contains breaking changes from v8.0.0-beta1 and v8.0.0-beta2.
+  Check out the explanation of these breaking changes below in order to migrate to v8.0.0.
+
+### v8.0.0-beta3
+
+- This version includes the fix from version 7.9.2.
+- `zod` version is 3.19.0.
+
+### v8.0.0-beta2
+
+- **Breaking**: removing the signature deprecated in v7.6.1.
+  - The argument of `EndpointsFactory::addMiddleware()` has to be the result of `createMiddleware()`.
+
+### v8.0.0-beta1
+
+- This version is based on v7.9.1.
+- **Breaking**: Only the following Node versions are supported:
+  - 14.17.0 and higher,
+  - 16.10.0 and higher
+  - 18.0.0 and higher.
+- `openapi3-ts` version is 3.0.2.
+- Supporting `jest` (optional peer dependency) version 29.x.
+
 ## Version 7
+
+### v7.9.2
+
+- Fixed issue #585 found and reported along with a suggested solution by [@rayzr522](https://github.com/rayzr522).
+  - In case you need to `throw` within an `Endpoint`'s handler or a `Middleware`, consider
+    [the best practice](https://eslint.org/docs/latest/rules/no-throw-literal) of only
+    throwing an `Error` or a _descendant_ that extends the `Error`.
+  - You can also `import { createHttpError } from "express-zod-api"` and use it for that purpose.
+  - However, this version fixes the issue caused by throwing something else.
+  - In this case that entity will be stringified into a `.message` of `Error`.
+  - The issue manifested itself as a positive API response without data.
+
+```typescript
+// reproduction example
+const myEndpoint = defaultEndpointsFactory.build({
+  method: "get",
+  input: z.object({}),
+  output: z.object({}),
+  handler: async () => {
+    throw "I'm not an Error";
+  },
+});
+```
+
+```json lines
+// response before:
+{"status":"success"}
+// response after:
+{"status":"error","error":{"message":"I'm not an Error"}}
+```
+
+### v7.9.1
+
+- Minor refactoring in order to support the recently released Typescript 4.8.2.
 
 ### v7.9.0